feat: migrate appLoader to ts
