+++ news-api/src/app/components/app/app.ts
@@ -2,17 +2,17 @@ import AppController from '../controller/controller';
 import { AppView } from '../view/appView';
 
 class App {
-    constructor() {
-        this.controller = new AppController();
-        this.view = new AppView();
-    }
+  constructor() {
+    this.controller = new AppController();
+    this.view = new AppView();
+  }
 
-    start() {
-        document
-            .querySelector('.sources')
-            .addEventListener('click', (e) => this.controller.getNews(e, (data) => this.view.drawNews(data)));
-        this.controller.getSources((data) => this.view.drawSources(data));
-    }
+  start() {
+    document
+      .querySelector('.sources')
+      .addEventListener('click', (e) => this.controller.getNews(e, (data) => this.view.drawNews(data)));
+    this.controller.getSources((data) => this.view.drawSources(data));
+  }
 }
 
 export default App;
